Initial commspace restructuring
