+++ core/src/Cabana_Halo.hpp
@@ -55,10 +55,13 @@ namespace Cabana
   ghost from is the unique owner of that data. Import is used in the context
   of the forward communication plan (the gather).
 */
-template <class MemorySpace, class BuildType = Export>
-class Halo : public CommunicationPlan<MemorySpace>
+template <class MemorySpace, class BuildType = Export,
+          class CommSpace = CommSpace::Mpi>
+class Halo : public CommunicationPlan<MemorySpace, CommSpace>
 {
   public:
+    using commspace_type = CommSpace;
+
     /*!
       \brief Neighbor and export rank constructor. Use this when you don't know
       who you will receiving from - only who you are sending to, but you already
@@ -104,14 +107,14 @@ class Halo : public CommunicationPlan<MemorySpace>
     Halo( MPI_Comm comm, const std::size_t num_local,
           const IdViewType& element_ids, const RankViewType& element_ranks,
           const std::vector<int>& neighbor_ranks )
-        : CommunicationPlan<MemorySpace>( comm )
+        : CommunicationPlan<MemorySpace, CommSpace>( comm )
         , _num_local( num_local )
     {
         if ( element_ids.size() != element_ranks.size() )
             throw std::runtime_error( "Cabana::Halo (export): ids and ranks "
                                       "views are different sizes!" );
 
-        auto neighbor_ids = this->createWithTopology(
+        auto neighbor_ids = this->createFromTopology(
             BuildType(), element_ranks, neighbor_ranks );
         this->createExportSteering( neighbor_ids, element_ranks, element_ids );
     }
@@ -152,19 +155,19 @@ class Halo : public CommunicationPlan<MemorySpace>
               std::enable_if_t<std::is_same<T, Export>::value, int> = 0>
     Halo( MPI_Comm comm, const std::size_t num_local,
           const IdViewType& element_ids, const RankViewType& element_ranks )
-        : CommunicationPlan<MemorySpace>( comm )
+        : CommunicationPlan<MemorySpace, CommSpace>( comm )
         , _num_local( num_local )
     {
         if ( element_ids.size() != element_ranks.size() )
-            throw std::runtime_error( "Cabana::Halo (export): ids and ranks "
+            throw std::runtime_error( "Cabana::Halo (import): ids and ranks "
                                       "views are different sizes!" );
 
         auto neighbor_ids =
-            this->createWithoutTopology( BuildType(), element_ranks );
+            this->createFromNoTopology( BuildType(), element_ranks );
         this->createExportSteering( neighbor_ids, element_ranks, element_ids );
     }
 
-    /*!
+     /*!
       \brief Neighbor and import rank constructor. Use this when you don't know
       who you will sending to - only who you are receiving from, but you already
       know which ranks neighbor each other (i.e. every rank already knows who
@@ -209,21 +212,21 @@ class Halo : public CommunicationPlan<MemorySpace>
     Halo( MPI_Comm comm, const std::size_t num_local,
           const IdViewType& element_ids, const RankViewType& element_ranks,
           const std::vector<int>& neighbor_ranks )
-        : CommunicationPlan<MemorySpace>( comm )
+        : CommunicationPlan<MemorySpace, CommSpace>( comm )
         , _num_local( num_local )
     {
         if ( element_ids.size() != element_ranks.size() )
             throw std::runtime_error( "Cabana::Halo (import): ids and ranks "
                                       "views are different sizes!" );
 
-        auto neighbor_ids_ranks_indices = this->createWithTopology(
+        auto neighbor_ids_ranks_indices = this->createFromTopology(
             BuildType(), element_ranks, element_ids, neighbor_ranks );
         this->createExportSteering( std::get<0>( neighbor_ids_ranks_indices ),
                                     std::get<1>( neighbor_ids_ranks_indices ),
                                     std::get<2>( neighbor_ids_ranks_indices ) );
     }
 
-    /*!
+        /*!
       \brief Import rank constructor. Use this when you don't know which ranks
       neighbor each other. (i.e. every rank does not already know who they will
       be exporting to and receiving from)
@@ -258,14 +261,14 @@ class Halo : public CommunicationPlan<MemorySpace>
               std::enable_if_t<std::is_same<T, Import>::value, int> = 0>
     Halo( MPI_Comm comm, const std::size_t num_local,
           const IdViewType& element_ids, const RankViewType& element_ranks )
-        : CommunicationPlan<MemorySpace>( comm )
+        : CommunicationPlan<MemorySpace, CommSpace>( comm )
         , _num_local( num_local )
     {
         if ( element_ids.size() != element_ranks.size() )
             throw std::runtime_error( "Cabana::Halo (import): ids and ranks "
                                       "views are different sizes!" );
 
-        auto neighbor_ids_ranks_indices = this->createWithoutTopology(
+        auto neighbor_ids_ranks_indices = this->createFromNoTopology(
             BuildType(), element_ranks, element_ids );
         this->createExportSteering( std::get<0>( neighbor_ids_ranks_indices ),
                                     std::get<1>( neighbor_ids_ranks_indices ),
@@ -275,7 +278,7 @@ class Halo : public CommunicationPlan<MemorySpace>
     /*!
       \brief Get the number of elements locally owned by this rank.
 
-      \return The number of elements locally owned by this rank.
+      \return THe number of elements locally owned by this rank.
     */
     std::size_t numLocal() const { return _num_local; }
 
@@ -390,91 +393,15 @@ class Gather<HaloType, AoSoAType,
     auto totalReceive() { return _halo.totalNumImport(); }
 
     /*!
-      \brief Perform the gather operation.
+    \brief Perform the gather operation.
     */
-    template <class ExecutionSpace>
-    void apply( ExecutionSpace )
-    {
-        Kokkos::Profiling::ScopedRegion region( "Cabana::Gather::apply" );
+    void apply() override { applyImpl( execution_space{}, commspace_type{} ); }
 
-        // Get the buffers and particle data (local copies for lambdas below).
-        auto send_buffer = this->getSendBuffer();
-        auto recv_buffer = this->getReceiveBuffer();
-        auto aosoa = this->getData();
+    template <class ExecutionSpace, class CommSpaceType>
+    std::enable_if_t<std::is_same<CommSpaceType, CommSpace::Mpi>::value, void>
+        applyImpl( ExecutionSpace, CommSpaceType );
 
-        // Get the steering vector for the sends.
-        auto steering = _halo.getExportSteering();
-        // Gather from the local data into a tuple-contiguous send buffer.
-        auto gather_send_buffer_func = KOKKOS_LAMBDA( const std::size_t i )
-        {
-            send_buffer( i ) = aosoa.getTuple( steering( i ) );
-        };
-        Kokkos::RangePolicy<ExecutionSpace> send_policy( 0, _send_size );
-        Kokkos::parallel_for( "Cabana::Gather::gather_send_buffer", send_policy,
-                              gather_send_buffer_func );
-        Kokkos::fence();
-
-        // The halo has it's own communication space so choose any mpi tag.
-        const int mpi_tag = 2345;
-
-        // Post non-blocking receives.
-        int num_n = _halo.numNeighbor();
-        std::vector<MPI_Request> requests( num_n );
-        std::pair<std::size_t, std::size_t> recv_range = { 0, 0 };
-        for ( int n = 0; n < num_n; ++n )
-        {
-            recv_range.second = recv_range.first + _halo.numImport( n );
-
-            auto recv_subview = Kokkos::subview( recv_buffer, recv_range );
-
-            MPI_Irecv( recv_subview.data(),
-                       recv_subview.size() * sizeof( data_type ), MPI_BYTE,
-                       _halo.neighborRank( n ), mpi_tag, _halo.comm(),
-                       &( requests[n] ) );
-
-            recv_range.first = recv_range.second;
-        }
-
-        // Do blocking sends.
-        std::pair<std::size_t, std::size_t> send_range = { 0, 0 };
-        for ( int n = 0; n < num_n; ++n )
-        {
-            send_range.second = send_range.first + _halo.numExport( n );
-
-            auto send_subview = Kokkos::subview( send_buffer, send_range );
-
-            MPI_Send( send_subview.data(),
-                      send_subview.size() * sizeof( data_type ), MPI_BYTE,
-                      _halo.neighborRank( n ), mpi_tag, _halo.comm() );
-
-            send_range.first = send_range.second;
-        }
-
-        // Wait on non-blocking receives.
-        std::vector<MPI_Status> status( num_n );
-        const int ec =
-            MPI_Waitall( requests.size(), requests.data(), status.data() );
-        if ( MPI_SUCCESS != ec )
-            throw std::logic_error(
-                "Cabana::Gather::apply (AoSoA): Failed MPI Communication" );
-
-        // Extract the receive buffer into the ghosted elements.
-        std::size_t num_local = _halo.numLocal();
-        auto extract_recv_buffer_func = KOKKOS_LAMBDA( const std::size_t i )
-        {
-            std::size_t ghost_idx = i + num_local;
-            aosoa.setTuple( ghost_idx, recv_buffer( i ) );
-        };
-        Kokkos::RangePolicy<ExecutionSpace> recv_policy( 0, _recv_size );
-        Kokkos::parallel_for( "Cabana::Gather::apply::extract_recv_buffer",
-                              recv_policy, extract_recv_buffer_func );
-        Kokkos::fence();
-
-        // Barrier before completing to ensure synchronization.
-        MPI_Barrier( _halo.comm() );
-    }
-
-    void apply() override { apply( execution_space{} ); }
+    // Future: Add applyImpl that is enabled for other CommSpaceType types.
 
     /*!
       \brief Reserve new buffers as needed and update the halo and AoSoA data.
@@ -576,110 +503,15 @@ class Gather<HaloType, SliceType,
     auto totalReceive() { return _halo.totalNumImport(); }
 
     /*!
-      \brief Perform the gather operation.
+    \brief Perform the gather operation.
     */
-    template <class ExecutionSpace>
-    void apply( ExecutionSpace )
-    {
-        Kokkos::Profiling::ScopedRegion region( "Cabana::Gather::apply" );
+    void apply() override { applyImpl( execution_space{}, commspace_type{} ); }
 
-        // Get the buffers (local copies for lambdas below).
-        auto send_buffer = this->getSendBuffer();
-        auto recv_buffer = this->getReceiveBuffer();
-        auto slice = this->getData();
+    template <class ExecutionSpace, class CommSpaceType>
+    std::enable_if_t<std::is_same<CommSpaceType, CommSpace::Mpi>::value, void>
+        applyImpl( ExecutionSpace, CommSpaceType );
 
-        // Get the number of components in the slice.
-        std::size_t num_comp = this->getSliceComponents();
-
-        // Get the raw slice data.
-        auto slice_data = slice.data();
-
-        // Get the steering vector for the sends.
-        auto steering = _halo.getExportSteering();
-
-        // Gather from the local data into a tuple-contiguous send buffer.
-        auto gather_send_buffer_func = KOKKOS_LAMBDA( const std::size_t i )
-        {
-            auto s = SliceType::index_type::s( steering( i ) );
-            auto a = SliceType::index_type::a( steering( i ) );
-            std::size_t slice_offset = s * slice.stride( 0 ) + a;
-            for ( std::size_t n = 0; n < num_comp; ++n )
-                send_buffer( i, n ) =
-                    slice_data[slice_offset + n * SliceType::vector_length];
-        };
-        Kokkos::RangePolicy<ExecutionSpace> send_policy( 0, _send_size );
-        Kokkos::parallel_for( "Cabana::Gather::gather_send_buffer", send_policy,
-                              gather_send_buffer_func );
-        Kokkos::fence();
-
-        // The halo has it's own communication space so choose any mpi tag.
-        const int mpi_tag = 2345;
-
-        // Post non-blocking receives.
-        int num_n = _halo.numNeighbor();
-        std::vector<MPI_Request> requests( num_n );
-        std::pair<std::size_t, std::size_t> recv_range = { 0, 0 };
-        for ( int n = 0; n < num_n; ++n )
-        {
-            recv_range.second = recv_range.first + _halo.numImport( n );
-
-            auto recv_subview =
-                Kokkos::subview( recv_buffer, recv_range, Kokkos::ALL );
-
-            MPI_Irecv( recv_subview.data(),
-                       recv_subview.size() * sizeof( data_type ), MPI_BYTE,
-                       _halo.neighborRank( n ), mpi_tag, _halo.comm(),
-                       &( requests[n] ) );
-
-            recv_range.first = recv_range.second;
-        }
-
-        // Do blocking sends.
-        std::pair<std::size_t, std::size_t> send_range = { 0, 0 };
-        for ( int n = 0; n < num_n; ++n )
-        {
-            send_range.second = send_range.first + _halo.numExport( n );
-
-            auto send_subview =
-                Kokkos::subview( send_buffer, send_range, Kokkos::ALL );
-
-            MPI_Send( send_subview.data(),
-                      send_subview.size() * sizeof( data_type ), MPI_BYTE,
-                      _halo.neighborRank( n ), mpi_tag, _halo.comm() );
-
-            send_range.first = send_range.second;
-        }
-
-        // Wait on non-blocking receives.
-        std::vector<MPI_Status> status( num_n );
-        const int ec =
-            MPI_Waitall( requests.size(), requests.data(), status.data() );
-        if ( MPI_SUCCESS != ec )
-            throw std::logic_error(
-                "Cabana::Gather::apply (Slice): Failed MPI Communication" );
-
-        // Extract the receive buffer into the ghosted elements.
-        std::size_t num_local = _halo.numLocal();
-        auto extract_recv_buffer_func = KOKKOS_LAMBDA( const std::size_t i )
-        {
-            std::size_t ghost_idx = i + num_local;
-            auto s = SliceType::index_type::s( ghost_idx );
-            auto a = SliceType::index_type::a( ghost_idx );
-            std::size_t slice_offset = s * slice.stride( 0 ) + a;
-            for ( std::size_t n = 0; n < num_comp; ++n )
-                slice_data[slice_offset + SliceType::vector_length * n] =
-                    recv_buffer( i, n );
-        };
-        Kokkos::RangePolicy<ExecutionSpace> recv_policy( 0, _recv_size );
-        Kokkos::parallel_for( "Cabana::Gather::extract_recv_buffer",
-                              recv_policy, extract_recv_buffer_func );
-        Kokkos::fence();
-
-        // Barrier before completing to ensure synchronization.
-        MPI_Barrier( _halo.comm() );
-    }
-
-    void apply() override { apply( execution_space{} ); }
+    // Future: Add applyImpl that is enabled for other CommSpaceType types.
 
     /*!
       \brief Reserve new buffers as needed and update the halo and slice data.
@@ -833,112 +665,13 @@ class Scatter
     /*!
       \brief Perform the scatter operation.
     */
-    template <class ExecutionSpace>
-    void apply( ExecutionSpace )
-    {
-        Kokkos::Profiling::ScopedRegion region( "Cabana::Scatter::apply" );
+    void apply() override { applyImpl( execution_space{}, commspace_type{} ); }
 
-        // Get the buffers (local copies for lambdas below).
-        auto send_buffer = this->getSendBuffer();
-        auto recv_buffer = this->getReceiveBuffer();
-        auto slice = this->getData();
+    template <class ExecutionSpace, class CommSpaceType>
+    std::enable_if_t<std::is_same<CommSpaceType, CommSpace::Mpi>::value, void>
+        applyImpl( ExecutionSpace, CommSpaceType );
 
-        // Get the number of components in the slice.
-        std::size_t num_comp = this->getSliceComponents();
-
-        // Get the raw slice data. Wrap in a 1D Kokkos View so we can unroll the
-        // components of each slice element.
-        Kokkos::View<data_type*, memory_space,
-                     Kokkos::MemoryTraits<Kokkos::Unmanaged>>
-            slice_data( slice.data(), slice.numSoA() * slice.stride( 0 ) );
-
-        // Extract the send buffer from the ghosted elements.
-        std::size_t num_local = _halo.numLocal();
-        auto extract_send_buffer_func = KOKKOS_LAMBDA( const std::size_t i )
-        {
-            std::size_t ghost_idx = i + num_local;
-            auto s = SliceType::index_type::s( ghost_idx );
-            auto a = SliceType::index_type::a( ghost_idx );
-            std::size_t slice_offset = s * slice.stride( 0 ) + a;
-            for ( std::size_t n = 0; n < num_comp; ++n )
-                send_buffer( i, n ) =
-                    slice_data( slice_offset + SliceType::vector_length * n );
-        };
-        Kokkos::RangePolicy<ExecutionSpace> send_policy( 0, _send_size );
-        Kokkos::parallel_for( "Cabana::Scatter::apply::extract_send_buffer",
-                              send_policy, extract_send_buffer_func );
-        Kokkos::fence();
-
-        // The halo has it's own communication space so choose any mpi tag.
-        const int mpi_tag = 2345;
-
-        // Post non-blocking receives.
-        int num_n = _halo.numNeighbor();
-        std::vector<MPI_Request> requests( num_n );
-        std::pair<std::size_t, std::size_t> recv_range = { 0, 0 };
-        for ( int n = 0; n < num_n; ++n )
-        {
-            recv_range.second = recv_range.first + _halo.numExport( n );
-
-            auto recv_subview =
-                Kokkos::subview( recv_buffer, recv_range, Kokkos::ALL );
-
-            MPI_Irecv( recv_subview.data(),
-                       recv_subview.size() * sizeof( data_type ), MPI_BYTE,
-                       _halo.neighborRank( n ), mpi_tag, _halo.comm(),
-                       &( requests[n] ) );
-
-            recv_range.first = recv_range.second;
-        }
-
-        // Do blocking sends.
-        std::pair<std::size_t, std::size_t> send_range = { 0, 0 };
-        for ( int n = 0; n < num_n; ++n )
-        {
-            send_range.second = send_range.first + _halo.numImport( n );
-
-            auto send_subview =
-                Kokkos::subview( send_buffer, send_range, Kokkos::ALL );
-
-            MPI_Send( send_subview.data(),
-                      send_subview.size() * sizeof( data_type ), MPI_BYTE,
-                      _halo.neighborRank( n ), mpi_tag, _halo.comm() );
-
-            send_range.first = send_range.second;
-        }
-
-        // Wait on non-blocking receives.
-        std::vector<MPI_Status> status( num_n );
-        const int ec =
-            MPI_Waitall( requests.size(), requests.data(), status.data() );
-        if ( MPI_SUCCESS != ec )
-            throw std::logic_error( "Cabana::Scatter::apply (Slice): "
-                                    "Failed MPI Communication" );
-
-        // Get the steering vector for the sends.
-        auto steering = _halo.getExportSteering();
-
-        // Scatter the ghosts in the receive buffer into the local values.
-        auto scatter_recv_buffer_func = KOKKOS_LAMBDA( const std::size_t i )
-        {
-            auto s = SliceType::index_type::s( steering( i ) );
-            auto a = SliceType::index_type::a( steering( i ) );
-            std::size_t slice_offset = s * slice.stride( 0 ) + a;
-            for ( std::size_t n = 0; n < num_comp; ++n )
-                Kokkos::atomic_add(
-                    &slice_data( slice_offset + SliceType::vector_length * n ),
-                    recv_buffer( i, n ) );
-        };
-        Kokkos::RangePolicy<ExecutionSpace> recv_policy( 0, _recv_size );
-        Kokkos::parallel_for( "Cabana::Scatter::apply::scatter_recv_buffer",
-                              recv_policy, scatter_recv_buffer_func );
-        Kokkos::fence();
-
-        // Barrier before completing to ensure synchronization.
-        MPI_Barrier( _halo.comm() );
-    }
-
-    void apply() override { apply( execution_space{} ); }
+    // Future: Add applyImpl that is enabled for other CommSpaceType types.
 
     /*!
       \brief Reserve new buffers as needed and update the halo and slice data.
@@ -983,6 +716,16 @@ class Scatter
     using base_type::_recv_size;
     using base_type::_send_size;
 };
+
+} // end namespace Cabana
+
+// Include communication backends from what is enabled in CMake.
+#ifdef Cabana_ENABLE_MPI
+#include <impl/Cabana_Halo_Mpi.hpp>
+#endif // Enable MPI
+
+namespace Cabana
+{
 
 /*!
   \brief Create the scatter.
