+++ core/unit_test/tstCommunicationPlan.hpp
@@ -10,7 +10,7 @@
  ****************************************************************************/
 
 #include <Cabana_AoSoA.hpp>
-#include <Cabana_CommunicationPlan.hpp>
+#include <Cabana_CommunicationPlanBase.hpp>
 #include <Cabana_DeepCopy.hpp>
 
 #include <Kokkos_Core.hpp>
@@ -975,54 +975,99 @@ void test11( const bool use_topology )
 //---------------------------------------------------------------------------//
 // RUN TESTS
 //---------------------------------------------------------------------------//
+template <typename TEST_COMMSPACE>
+class CommunicationPlanTypedTest : public ::testing::Test
+{
+  public:
+    using CommSpaceType = TEST_COMMSPACE;
+};
+
+using CommSpaceTypes = ::testing::Types<Cabana::CommSpace::Mpi
+// When added, place additional CommSpace types here to test
+                                        >;
+// Export tests
+YPED_TEST_SUITE( CommunicationPlanTypedTest, CommSpaceTypes );
 
 // Export tests
-TEST( CommPlan, Test1 ) { test1( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test1 ) { test1<TypeParam>( true ); }
 
-TEST( CommPlan, Test2 ) { test2( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test2 ) { test2<TypeParam>( true ); }
 
-TEST( CommPlan, Test3 ) { test3( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test3 ) { test3<TypeParam>( true ); }
 
-TEST( CommPlan, Test4 ) { test4( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test4 ) { test4<TypeParam>( true ); }
 
-TEST( CommPlan, Test5 ) { test5( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test5 ) { test5<TypeParam>( true ); }
 
-TEST( CommPlan, Test6 ) { test6( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test6 ) { test6<TypeParam>( true ); }
 
-TEST( CommPlan, Test7 ) { test7( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test7 ) { test7<TypeParam>( true ); }
 
-TEST( CommPlan, Test1NoTopo ) { test1( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test1NoTopo )
+{
+    test1<TypeParam>( false );
+}
 
-TEST( CommPlan, Test2NoTopo ) { test2( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test2NoTopo )
+{
+    test2<TypeParam>( false );
+}
 
-TEST( CommPlan, Test3NoTopo ) { test3( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test3NoTopo )
+{
+    test3<TypeParam>( false );
+}
 
-TEST( CommPlan, Test4NoTopo ) { test4( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test4NoTopo )
+{
+    test4<TypeParam>( false );
+}
 
-TEST( CommPlan, Test5NoTopo ) { test5( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test5NoTopo )
+{
+    test5<TypeParam>( false );
+}
 
-TEST( CommPlan, Test6NoTopo ) { test6( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test6NoTopo )
+{
+    test6<TypeParam>( false );
+}
 
-TEST( CommPlan, Test7NoTopo ) { test7( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test7NoTopo )
+{
+    test7<TypeParam>( false );
+}
 
-TEST( CommPlan, TestTopology ) { testTopology(); }
+TEST( CommSpace, TestTopology ) { testTopology(); }
 
 // Import tests
-TEST( CommPlan, Test8 ) { test8( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test8 ) { test8<TypeParam>( true ); }
 
-TEST( CommPlan, Test9 ) { test9( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test9 ) { test9<TypeParam>( true ); }
 
-TEST( CommPlan, Test10 ) { test10( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test10 ) { test10<TypeParam>( true ); }
 
-TEST( CommPlan, Test11 ) { test11( true ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test11 ) { test11<TypeParam>( true ); }
 
-TEST( CommPlan, Test8NoTopo ) { test8( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test8NoTopo )
+{
+    test8<TypeParam>( false );
+}
 
-TEST( CommPlan, Test9NoTopo ) { test9( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test9NoTopo )
+{
+    test9<TypeParam>( false );
+}
 
-TEST( CommPlan, Test10NoTopo ) { test10( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test10NoTopo )
+{
+    test10<TypeParam>( false );
+}
 
-TEST( CommPlan, Test11NoTopo ) { test11( false ); }
+TYPED_TEST( CommunicationPlanTypedTest, Test11NoTopo )
+{
+    test11<TypeParam>( false );
+}
 
 //---------------------------------------------------------------------------//
